+++ rust/serving/src/lib.rs
@@ -6,8 +6,14 @@ use axum_server::tls_rustls::RustlsConfig;
 use tracing::info;
 
 pub use self::error::{Error, Result};
+use crate::app::orchestrator::OrchestratorState as CallbackState;
 use crate::app::start_main_server;
+use crate::app::store::cbstore::jetstreamstore::JetStreamCallbackStore;
+use crate::app::store::datastore::jetstream::JetStreamDataStore;
+use crate::app::store::datastore::user_defined::UserDefinedStore;
+use crate::app::tracker::MessageGraph;
 use crate::config::generate_certs;
+use crate::config::StoreType;
 use crate::metrics::start_https_metrics_server;
 use app::orchestrator::OrchestratorState;
 
@@ -24,9 +30,7 @@ mod metrics;
 mod pipeline;
 
 pub mod source;
-pub use source::start;
 pub use source::Message;
-use source::MessageWrapper;
 
 use crate::app::store::cbstore::CallbackStore;
 use crate::app::store::datastore::DataStore;
@@ -41,7 +45,7 @@ pub(crate) struct AppState<T, U> {
     pub(crate) orchestrator_state: OrchestratorState<T, U>,
 }
 
-pub(crate) async fn serve<T, U>(
+async fn serve<T, U>(
     app: AppState<T, U>,
 ) -> std::result::Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>
 where
@@ -78,4 +82,43 @@ async fn flatten<T>(handle: tokio::task::JoinHandle<Result<T>>) -> Result<T> {
         Ok(Err(err)) => Err(err),
         Err(err) => Err(Error::Other(format!("Spawning the server: {err:?}"))),
     }
+}
+
+pub async fn start(js_context: Context, settings: Arc<Settings>) -> Result<()> {
+    // create a callback store for tracking
+    let callback_store =
+        JetStreamCallbackStore::new(js_context.clone(), &settings.js_store).await?;
+    // Create the message graph from the pipeline spec and the redis store
+    let msg_graph = MessageGraph::from_pipeline(&settings.pipeline_spec).map_err(|e| {
+        Error::InitError(format!(
+            "Creating message graph from pipeline spec: {:?}",
+            e
+        ))
+    })?;
+
+    // Create a redis store to store the callbacks and the custom responses
+    match &settings.store_type {
+        StoreType::Nats => {
+            let nats_store =
+                JetStreamDataStore::new(js_context.clone(), &settings.js_store).await?;
+            let callback_state = CallbackState::new(msg_graph, nats_store, callback_store).await?;
+            let app = crate::AppState {
+                js_context,
+                settings,
+                orchestrator_state: callback_state,
+            };
+            crate::serve(app).await.unwrap();
+        }
+        StoreType::UserDefined(ud_config) => {
+            let ud_store = UserDefinedStore::new(ud_config.clone()).await?;
+            let callback_state = CallbackState::new(msg_graph, ud_store, callback_store).await?;
+            let app = crate::AppState {
+                js_context,
+                settings,
+                orchestrator_state: callback_state,
+            };
+            crate::serve(app).await.unwrap();
+        }
+    }
+    Ok(())
 }
